+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Behandlingsresultat.tsx
@@ -19,7 +19,7 @@ import OvergangsordningAndelTabell from './OvergangsordningAndel/Overgangsordnin
 import TilkjentYtelseTidslinje from './TilkjentYtelseTidslinje';
 import { useBehandlingsresultat } from './useBehandlingsresultat';
 import { useBehandling } from '../../../../../context/behandlingContext/BehandlingContext';
-import { kompetanseFeilmeldingId } from '../../../../../context/Kompetanse/KompetanseSkjemaContext';
+import { kompetanseFeilmeldingId } from '../../../../../context/Kompetanse/useKompetansePeriodeSkjema';
 import { useTidslinje } from '../../../../../context/TidslinjeContext';
 import { utenlandskPeriodeBeløpFeilmeldingId } from '../../../../../context/UtenlandskPeriodeBeløp/UtenlandskPeriodeBeløpSkjemaContext';
 import { valutakursFeilmeldingId } from '../../../../../context/Valutakurs/ValutakursSkjemaContext';
